Added 'sky camera'.
Press E on the monitor on the interior to activate. Left click sets
coordinates to be gone to with the throttle also in the interior, right
click zooms and slows down camera movement for more precise locating.
The bad throttle is going to be fixed in the next update, I think.