Added self-repair lever
Also added atom accelerator which spins during flight/teleporting and
can change direction by pressing E on it. Monitor also moved to new
entity system. Self-repair lever requires everyone to leave the tardis
before it activates, and no-one can go back in while it's repairing, the
light on the top will flash, there will be a noise emitted, and a
message will appear in chat to occupants who were in the tardis at time
of lever activation. It can also be deactivated before everyone leaves
the tardis to cancel self-repairs.