Added programmable flight mode.
Also wibbly lever now controls door lock, all the pully stick things
around the console are now animated, but most do not have a function
apart from the one that is the time-rotor handbrake. There's now a big
lever that acts as a fast return protocol to wherever the tardis last
departed from. The programmable flight mode has a sweet GUI to go with
it, and it saves data to the client, meaning you can permanently save
positions and load them up, individually for each map whenever you like.
Advanced mode is not yet available for programmable flight mode, so it
will work the same as it does in simple mode (advanced mode unchecked in
options), by just pressing the dematerialise button in the GUI.