Teleport onto walls/ceilings
Managed to get HitNormal working so you can now teleport onto walls or
even ceilings with the sonic or by clicking in flightmode, as per usual,
I'd recommend using the physbrake though during demat. Turning the power
back on now restores certain states, such as flightmode, lock and
physlock. The power is by default on for those who cannot get to the
interior to turn it on.