Finished simple long-flight.
To use the new longflight, turn it on in the admin options section of
tardis menu. Then dematerialise normally, but instead of reappearing
straight away at the new location it will simply disappear entirely. If
you're in third person view, click to initiate materialisation. If
you're in the interior, pull the throttle to initiate materialisation.
If you're somehow outside the tardis completely and you have a linked
sonic screwdriver, press R to initiate materialisation. At this time
advanced mode does not affect long flight, but it will in future. Other
changes: Green light stays on in the interior even when powered off.
Pressing other things during advanced start sequence no longer break it.
Added warning message when trying to use the isomorphic security system
in singleplayer. Added flight sounds when in the vortex using long
flight.