Player:GetPos() now returns position of tardis while inside.
Player:GetRealPos() is used to get the real location of the player,
which is sometimes required. This feature is slightly experimental so do
let me know if things start breaking and stuff. Also, Horizon support is
improved slightly, and fixed a lua error/bug when a player in the
skycamera left the server.