50th Anniversary
Removed leftover debug message from function ENT:SetTrackingEnt. Added
the Gramophone: This will allow you to stream music from my very own web
server from Doctor Who to everyone who is in the interior. Music
currently included is composed of the main themes, all doctors themes,
and all companion themes (2005-present). To activate the Gramophone
simply press E on it and a GUI will appear. Music can be lowered in
volume or turned off entirely from the new slider and tickbox in the
options. Fixed the light not always flashing when repairs are completed.
The volume of looping sounds (interior idle/flight loop) were decreased
to reduce annoyance, and the flight loop now changes pitch based on the
TARDIS speed if you're inside it, or based on the doppler effect if not.
The range of the exterior flight sound has been increased quite
dramatically, but not too much to induce annoyance. To further reduce
possible annoyance, a slider has been added to adjust exterior flight
sound volume. Repair times decreased by half, so a full repair now takes
50 seconds instead of 100. Spinmode E2 function added and chat command
added to official E2. Smoke now emits from the exterior when health is
at or below 20%, whereas previously it would only emit at 0% health.
Tracking should now follow a little more smoothly. The 'Scanner' (big
window to the left of the entrance) is now fully functional, It will
display just outside the doors of the TARDIS. Note also that this
feature could possibly be a bit buggy so let me know if somethings wrong
with it.