50th Anniversary fixes and more
This update aims to fix issues that arised in the 50th Anniversary
update. Fixed 'halo's that appear on the console parts, also added an
option to disable halos completely (for a possible framerate boost).
Added ability to hear interior music while viewing in third person, on
by default but can be turned off in options (Flightmode music). Added
Abigail's Song to the Gramophone. Added a button to the gramophone GUI
which allows the player to select any music from a website, as long as
it's in one of the required formats (MP3, MP2, MP1, OGG, WAV, AIFF).
Changed the gramophone error message to be a bit more informative. Added
an option to disable the scanner for a framerate boost. Fixed a bug
where setting a destination after dematerialising in long flight while
exploded would result in rematerialising in the set location rather than
a random one. Fixed the fire and smoke being visible after
dematerialising.