Added Hostile Action Displacement System
The Hostile Action Displacement System (HADS) has been properly
implemented now. Taking up one of the two remaining 'unused levers', the
HADS will automatically (and quickly) dematerialise the box and put it
into longflight if it comes under danger (damage/water). To help bring
it back if you wern't inside of the box at the time, the sonic can now
set the remat location for longflight and an E2 function
(entity:tardisSetDestination) was also added to do the same. E2
functions were also created to toggle HADS/check if it's on
(entity:tardisHADS/entity:tardisIsHADS respectively), as well as a
function to check if it's in the vortex (entity:tardisInVortex()). A new
dematerialisation type called 'quickdemat' internally will dematerialise
the TARDIS 1.75 times faster than usual - and is currently only used for
HADS. Added several new songs to the Gramophone. Changed the interior
door to a physical entity, allowing for future improvements and stuff.
The exterior flight sound (from a viewpoint other than tardis third
person) now includes a higher pitch for going faster as well as doppler
effect. Fixed the music not playing properly when using the skycamera.