Updated models
Repositioned and resized portals on exterior and interior. Adjusted
interior to work with new portal position. Exterior now has black walls
on the inside so you can't see through it by looking at a very high
angle very close to it. Added cool colour changing effect to the screens
because why not.