Restructured to support Doors addon
I've moved all of the bigger-on-the-inside stuff into a separate addon
named Doors and converted this to use it. I've also fixed a few
important bugs including an important bug relating to base entities,
which was crippling the hook system. I've also added a new gamemode
hook, 'SkinChanged' which is called whenever something changes it's skin
serverside. It's used to change the door skin whenever the main
exterior's skin is changed.

- Moved all bigger-on-the-inside code into separate addon, Doors
- Improved the fallback force for no gravity
- Enter/exit fallback angle is now calculated automatically
- Fixed base class issue with hook system
- Made the derma skin module clientside (was shared)
- Added generic 'stuckfilter' which excludes entities in it from stuck
check
- Added more restrictions against the interior and parts (fixes #118)
- Fixed the addon breaking when the clientside interior setting was
invalid (fixes #111)
- Door skins now change when exterior skin is changed (fixes #12)
- Interior can no longer be seen through scanner (fixes #108)