Various fixes and improvements
This update includes various bug fixes and a few enhancements. The music
player now uses my own custom API so it should work a lot better than
the previous one. I'm not sure how it's going to handle the load but
we'll find out. There's also improved support for transparent parts that
would sometimes disappear and support for chairs in the interior.

- Switched music player to use my API instead
- Fixed #140 (thirdperson error)
- Added PART.UseTransparencyFix to work around disappearing parts (fixes
#136)
- Fixed T.Interior.ExitDistance being locked to 600
- Fixed #134 (player handling error)
- Added T.Interior.Seats (seats in the interior, fixes #137)
- Added dev command tardis_getlocal <decimals> (prints entity local
pos/ang to interior)
- Fixed binds working while spawnmenu is open (fixes #135)
- Variables on parts are now accessible clientside as well
- Switched back to and fixed the newer exterior skin (fixes #138)