CSequences V.1 Welcome, to Era 3 :>
Advanced mode is back baby!
Using the latest technological advancements of 2019
we've managed the return of an old friend.

If you've ever felt lost in the Legacy TARDIS, fear not!
You're about to get lost again.

Please refer to the wiki page for how to create sequences
and add them to your own TARDIS.